+++ README.md
@@ -1,1 +1,10 @@
-# milenty.github.io
+# Сборка сайта
+## https://milenty.github.io
+### Учебный проект сборки сайта с помощью webpack
+## Разворачиваем проект из шаблона
+### Шаг 1 
+#### клонируем репозиторий к себе на компьютер
+### Шаг 2
+#### запускаем команду npm i
+### Шаг 3
+#### запускаем команду npm start или npm run dev
