+++ package.json
@@ -12,7 +12,7 @@
     "type": "git",
     "url": "git+https://github.com/milenty/milenty.github.io.git"
   },
-  "author": "",
+  "author": "Samorukova Maria",
   "license": "ISC",
   "bugs": {
     "url": "https://github.com/milenty/milenty.github.io/issues"
