+++ src/style.css
@@ -1,19 +1,6 @@
-@import url(./vendor/normalize.css);
-@import url(./vendor/fonts.css);
-@import url(./blocks/root/root.css);
-@import url(./blocks/header/header.css);
-@import url(./blocks/logo/logo.css);
-@import url(./blocks/profile/profile.css);
-@import url(./blocks/user-info/user-info.css);
-@import url(./blocks/button/button.css);
-@import url(./blocks/places-list/places-list.css);
-@import url(./blocks/place-card/place-card.css);
-@import url(./blocks/popup/popup.css);
-@import url(./blocks/popup-edit/popup-edit.css);
-@import url(./blocks/popup-img/popup-img.css);
+@import url(./pages/index.css);
 
 
-body {
-    background-color: #000;
-  }
+
+
